refactor: change names of component files, start them with Capital letters
