+++ src/components/CardGenerator.jsx
