+++ src/components/RitualSubjects.jsx
